refactor: change server action implementation, improve login UX♻️
This PR adds loading when the user submits the login form. To make this
work the underlying server action implementation needed to be
changed/updated which would then make the useFormStatus().pending to
work when submitting the form. Changing the server action implementation
actually broke the client side validation when submitting the form (Used
isValid hook from react-hook-form to make sure user cannot submit until
and unless the input fields are validated on the client side). And some
other ux enhances